Moves cli show_usage to help.py
